requirements.txt
# 核心依赖
torch>=1.12.0
transformers>=4.20.0
tokenizers>=0.12.0

# CKIP Transformers引擎
ckip-transformers>=0.4.0

# 数据处理
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.1.0

# 配置和工具
python-dotenv>=0.19.0
pyyaml>=6.0
click>=8.1.0

# 日志和监控
loguru>=0.6.0
tqdm>=4.64.0

# 测试
pytest>=7.0.0
pytest-cov>=4.0.0

# 开发工具
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0

# 可选：GPU支持
# torch-cuda>=1.12.0  # 如果需要GPU支持，取消注释 